requirements.txt
streamlit==1.37.1
pandas==2.2.2
openpyxl==3.1.5
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
pyyaml==6.0.2
cloudscraper==1.2.71
playwright==1.46.0
python-slugify==8.0.4
